restructure according to review
