+++ run_all_experiments.py
@@ -129,75 +129,6 @@ def prepare_experiment_targets(
   return experiment_configs
 
 
-def has_cache_build_script(project: str) -> bool:
-  """Checks if a project has cached fuzzer build script."""
-  cached_build_script = os.path.join('fuzzer_build_script', project)
-  return os.path.isfile(cached_build_script)
-
-
-def prepare_image_cache(project: str) -> bool:
-  """Prepares cached images of fuzzer build containers."""
-  # Only create a cached image if we have a post-build build script
-  if not has_cache_build_script(project):
-    logger.info('No cached script for %s', project)
-    return False
-  logger.info('%s has a cached build script', project)
-
-  cached_container_name = oss_fuzz_checkout.get_project_cache_name(project)
-  adjusted_env = os.environ | {
-      'OSS_FUZZ_SAVE_CONTAINERS_NAME': cached_container_name
-  }
-
-  logger.info('Creating a cached images')
-  for sanitizer in ['address', 'coverage']:
-    # Create cached image by building using OSS-Fuzz with set variable
-    command = [
-        'python3', 'infra/helper.py', 'build_fuzzers', project, '--sanitizer',
-        sanitizer
-    ]
-    try:
-      sp.run(command,
-             cwd=oss_fuzz_checkout.OSS_FUZZ_DIR,
-             env=adjusted_env,
-             check=True)
-    except sp.CalledProcessError:
-      logger.info('Failed to build fuzzer for %s.', project)
-      return False
-
-    # Commit the container to an image
-    cached_image_name = oss_fuzz_checkout.get_project_cache_image_name(
-        project, sanitizer)
-
-    command = ['docker', 'commit', cached_container_name, cached_image_name]
-    try:
-      sp.run(command, check=True)
-    except sp.CalledProcessError:
-      logger.info('Could not rename image.')
-      return False
-    logger.info('Created cached image %s', cached_image_name)
-
-    # Delete the container we created
-    command = ['docker', 'container', 'rm', cached_container_name]
-    try:
-      sp.run(command, check=True)
-    except sp.CalledProcessError:
-      logger.info('Could not rename image.')
-  return True
-
-
-def prepare_cached_images(
-    experiment_targets: list[benchmarklib.Benchmark]) -> None:
-  """Builds cached Docker images for a set of targets."""
-  all_projects = set()
-  for benchmark in experiment_targets:
-    all_projects.add(benchmark.project)
-
-  logger.info('Preparing cache for %d projects', len(all_projects))
-
-  for project in all_projects:
-    prepare_image_cache(project)
-
-
 def run_experiments(benchmark: benchmarklib.Benchmark,
                     args: argparse.Namespace) -> Result:
   """Runs an experiment based on the |benchmark| config."""
@@ -441,7 +372,7 @@ def main():
   experiment_results = []
 
   if oss_fuzz_checkout.ENABLE_CACHING:
-    prepare_cached_images(experiment_targets)
+    oss_fuzz_checkout.prepare_cached_images(experiment_targets)
 
   logger.info(f'Running %s experiment(s) in parallels of %s.',
               len(experiment_targets), str(NUM_EXP))
